requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
 flask-migrate==4.0.7
flask-cors==4.0.1
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
opencv-python==4.9.0.80
face-recognition==1.3.0
 dlib==19.24.4
 numpy==1.26.4
Pillow==10.3.0
gunicorn==22.0.0
